feat(blazorui): add URL match feature to BitNav #11225 (#11226)
