+++ src/BlazorUI/Bit.BlazorUI.Extras/Components/NavPanel/BitNavPanel.razor
@@ -127,6 +127,7 @@
                         Items="_filteredNavItems"
                         ItemTemplate="ItemTemplate"
                         ItemTemplateRenderMode="ItemTemplateRenderMode"
+                        Match="NavMatch"
                         Mode="NavMode"
                         NoCollapse="NoCollapse"
                         OnItemClick="(TItem item) => HandleNavItemClick(item)"
