+++ src/BlazorUI/Bit.BlazorUI/Components/Navs/Nav/BitNav.razor.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Components.Routing;
+﻿using System.Text.RegularExpressions;
+using Microsoft.AspNetCore.Components.Routing;
 
 namespace Bit.BlazorUI;
 
@@ -15,176 +16,6 @@ public partial class BitNav<TItem> : BitComponentBase where TItem : class
 
 
     [Inject] private NavigationManager _navigationManager { get; set; } = default!;
-
-
-
-    /// <summary>
-    /// The accent color of the nav.
-    /// </summary>
-    [Parameter, ResetClassBuilder]
-    public BitColor? Accent { get; set; }
-
-    /// <summary>
-    /// Expands all items on first render.
-    /// </summary>
-    [Parameter] public bool AllExpanded { get; set; }
-
-    /// <summary>
-    /// The custom icon name of the chevron-down element of each nav item.
-    /// </summary>
-    [Parameter] public string? ChevronDownIcon { get; set; }
-
-    /// <summary>
-    /// Items to render as children.
-    /// </summary>
-    [Parameter]
-    [CallOnSet(nameof(OnSetParameters))]
-    public RenderFragment? ChildContent { get; set; }
-
-    /// <summary>
-    /// Custom CSS classes for different parts of the nav.
-    /// </summary>
-    [Parameter] public BitNavClassStyles? Classes { get; set; }
-
-    /// <summary>
-    /// The general color of the nav that is only used for colored parts like icons.
-    /// </summary>
-    [Parameter, ResetClassBuilder]
-    public BitColor? Color { get; set; }
-
-    /// <summary>
-    /// The initially selected item in manual mode.
-    /// </summary>
-    [Parameter] public TItem? DefaultSelectedItem { get; set; }
-
-    /// <summary>
-    /// Renders the nav in a width to only fit its content.
-    /// </summary>
-    [Parameter, ResetClassBuilder]
-    public bool FitWidth { get; set; }
-
-    /// <summary>
-    /// Renders the nav in full width of its container element.
-    /// </summary>
-    [Parameter, ResetClassBuilder]
-    public bool FullWidth { get; set; }
-
-    /// <summary>
-    /// Used to customize how content inside the group header is rendered.
-    /// </summary>
-    [Parameter] public RenderFragment<TItem>? HeaderTemplate { get; set; }
-
-    /// <summary>
-    /// The render mode of the custom HeaderTemplate.
-    /// </summary>
-    [Parameter] public BitNavItemTemplateRenderMode HeaderTemplateRenderMode { get; set; }
-
-    /// <summary>
-    /// Only renders the icon of each nav item.
-    /// </summary>
-    [Parameter, ResetClassBuilder]
-    public bool IconOnly { get; set; }
-
-    /// <summary>
-    /// The indentation padding in px for items without children (compensation space for chevron icon).
-    /// </summary>
-    [Parameter] public int IndentPadding { get; set; } = 27;
-
-    /// <summary>
-    /// The indentation padding in px for items in reversed mode.
-    /// </summary>
-    [Parameter] public int IndentReversedPadding { get; set; } = 4;
-
-    /// <summary>
-    /// The indentation value in px for each level of depth of child item.
-    /// </summary>
-    [Parameter] public int IndentValue { get; set; } = 16;
-
-    /// <summary>
-    /// A collection of items to display in the BitNav component.
-    /// </summary>
-    [Parameter]
-    [CallOnSet(nameof(OnSetParameters))]
-    public IList<TItem> Items { get; set; } = [];
-
-    /// <summary>
-    /// Used to customize how content inside the item is rendered.
-    /// </summary>
-    [Parameter] public RenderFragment<TItem>? ItemTemplate { get; set; }
-
-    /// <summary>
-    /// The render mode of the custom ItemTemplate.
-    /// </summary>
-    [Parameter] public BitNavItemTemplateRenderMode ItemTemplateRenderMode { get; set; }
-
-    /// <summary>
-    /// Determines how the navigation will be handled.
-    /// </summary>
-    [Parameter]
-    [CallOnSet(nameof(OnSetMode))]
-    public BitNavMode Mode { get; set; }
-
-    /// <summary>
-    /// Names and selectors of the custom input type properties.
-    /// </summary>
-    [Parameter] public BitNavNameSelectors<TItem>? NameSelectors { get; set; }
-
-    /// <summary>
-    /// Hides all collapse/expand buttons and remove their spaces at the start of each node.
-    /// </summary>
-    [Parameter] public bool NoCollapse { get; set; }
-
-    /// <summary>
-    /// Callback invoked when an item is clicked.
-    /// </summary>
-    [Parameter] public EventCallback<TItem> OnItemClick { get; set; }
-
-    /// <summary>
-    /// Callback invoked when a group header is clicked and Expanded or Collapse.
-    /// </summary>
-    [Parameter] public EventCallback<TItem> OnItemToggle { get; set; }
-
-    /// <summary>
-    /// Callback invoked when an item is selected.
-    /// </summary>
-    [Parameter] public EventCallback<TItem> OnSelectItem { get; set; }
-
-    /// <summary>
-    /// Alias of ChildContent.
-    /// </summary>
-    [Parameter] public RenderFragment? Options { get; set; }
-
-    /// <summary>
-    /// The way to render nav items.
-    /// </summary>
-    [Parameter] public BitNavRenderType RenderType { get; set; }
-
-    /// <summary>
-    /// Enables recalling the select events when the same item is selected.
-    /// </summary>
-    [Parameter] public bool Reselectable { get; set; }
-
-    /// <summary>
-    /// Reverses the location of the expander chevron.
-    /// </summary>
-    [Parameter] public bool ReversedChevron { get; set; }
-
-    /// <summary>
-    /// Selected item to show in the BitNav.
-    /// </summary>
-    [Parameter, TwoWayBound]
-    [CallOnSet(nameof(OnSetSelectedItem))]
-    public TItem? SelectedItem { get; set; }
-
-    /// <summary>
-    /// Enables the single-expand mode in the BitNav.
-    /// </summary>
-    [Parameter] public bool SingleExpand { get; set; }
-
-    /// <summary>
-    /// Custom CSS styles for different parts of the BitNav component.
-    /// </summary>
-    [Parameter] public BitNavClassStyles? Styles { get; set; }
 
 
 
@@ -358,23 +189,6 @@ public partial class BitNav<TItem> : BitComponentBase where TItem : class
 
 
 
-    internal void SetIsExpanded(TItem item, bool value)
-    {
-        if (item is BitNavItem navItem)
-        {
-            navItem.IsExpanded = value;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            navOption.IsExpanded = value;
-        }
-
-        if (NameSelectors is null) return;
-
-        item.SetValueToProperty(NameSelectors.IsExpanded.Name, value);
-    }
-    
     internal void SetItemExpanded(TItem item, bool value)
     {
         var isExpanded = GetIsExpanded(item);
@@ -419,490 +233,6 @@ public partial class BitNav<TItem> : BitComponentBase where TItem : class
         StateHasChanged();
     }
 
-    internal BitNavAriaCurrent GetAriaCurrent(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.AriaCurrent;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.AriaCurrent;
-        }
-
-        if (NameSelectors is null) return BitNavAriaCurrent.Page;
-
-        if (NameSelectors.AriaCurrent.Selector is not null)
-        {
-            return NameSelectors.AriaCurrent.Selector!(item) ?? BitNavAriaCurrent.Page;
-        }
-
-        return item.GetValueFromProperty(NameSelectors.AriaCurrent.Name, BitNavAriaCurrent.Page);
-    }
-
-    internal string? GetAriaLabel(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.AriaLabel;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.AriaLabel;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.AriaLabel.Selector is not null)
-        {
-            return NameSelectors.AriaLabel.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.AriaLabel.Name);
-    }
-
-    internal string? GetClass(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Class;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Class;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Class.Selector is not null)
-        {
-            return NameSelectors.Class.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Class.Name);
-    }
-
-    internal List<TItem> GetChildItems(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.ChildItems as List<TItem> ?? [];
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.ChildItems as List<TItem> ?? [];
-        }
-
-        if (NameSelectors is null) return [];
-
-        if (NameSelectors.ChildItems.Selector is not null)
-        {
-            return NameSelectors.ChildItems.Selector!(item) ?? [];
-        }
-
-        return item.GetValueFromProperty<List<TItem>>(NameSelectors.ChildItems.Name, [])!;
-    }
-
-    internal string? GetCollapseAriaLabel(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.CollapseAriaLabel;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.CollapseAriaLabel;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.CollapseAriaLabel.Selector is not null)
-        {
-            return NameSelectors.CollapseAriaLabel.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.CollapseAriaLabel.Name);
-    }
-
-    internal object? GetData(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Data;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Data;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Data.Selector is not null)
-        {
-            return NameSelectors.Data.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Data.Name);
-    }
-
-    internal string? GetDescription(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Description;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Description;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Description.Selector is not null)
-        {
-            return NameSelectors.Description.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Description.Name);
-    }
-
-    internal string? GetExpandAriaLabel(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.ExpandAriaLabel;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.ExpandAriaLabel;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.ExpandAriaLabel.Selector is not null)
-        {
-            return NameSelectors.ExpandAriaLabel.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.ExpandAriaLabel.Name);
-    }
-
-    internal bool GetForceAnchor(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.ForceAnchor;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.ForceAnchor;
-        }
-
-        if (NameSelectors is null) return false;
-
-        if (NameSelectors.ForceAnchor.Selector is not null)
-        {
-            return NameSelectors.ForceAnchor.Selector!(item) ?? false;
-        }
-
-        return item.GetValueFromProperty(NameSelectors.ForceAnchor.Name, false);
-    }
-
-    internal string? GetIconName(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.IconName;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.IconName;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.IconName.Selector is not null)
-        {
-            return NameSelectors.IconName.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.IconName.Name);
-    }
-
-    internal bool GetIsEnabled(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.IsEnabled;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.IsEnabled;
-        }
-
-        if (NameSelectors is null) return true;
-
-        if (NameSelectors.IsEnabled.Selector is not null)
-        {
-            return NameSelectors.IsEnabled.Selector!(item) ?? true;
-        }
-
-        return item.GetValueFromProperty(NameSelectors.IsEnabled.Name, true);
-    }
-
-    internal bool? GetIsExpanded(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.IsExpanded;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.IsExpanded;
-        }
-
-        if (NameSelectors is null) return false;
-
-        if (NameSelectors.IsExpanded.Selector is not null)
-        {
-            return NameSelectors.IsExpanded.Selector!(item) ?? false;
-        }
-
-        return item.GetValueFromProperty(NameSelectors.IsExpanded.Name, false);
-    }
-
-    internal bool GetIsSeparator(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.IsSeparator;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.IsSeparator;
-        }
-
-        if (NameSelectors is null) return false;
-
-        if (NameSelectors.IsSeparator.Selector is not null)
-        {
-            return NameSelectors.IsSeparator.Selector!(item) ?? false;
-        }
-
-        return item.GetValueFromProperty(NameSelectors.IsSeparator.Name, false);
-    }
-
-    internal string? GetKey(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Key;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Key;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Key.Selector is not null)
-        {
-            return NameSelectors.Key.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Key.Name);
-    }
-
-    internal string? GetStyle(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Style;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Style;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Style.Selector is not null)
-        {
-            return NameSelectors.Style.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Style.Name);
-    }
-
-    internal string? GetTarget(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Target;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Target;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Target.Selector is not null)
-        {
-            return NameSelectors.Target.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Target.Name);
-    }
-
-    internal RenderFragment<TItem>? GetTemplate(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Template as RenderFragment<TItem>;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Template as RenderFragment<TItem>;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Template.Selector is not null)
-        {
-            return NameSelectors.Template.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<RenderFragment<TItem>?>(NameSelectors.Template.Name);
-    }
-
-    internal BitNavItemTemplateRenderMode GetTemplateRenderMode(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.TemplateRenderMode;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.TemplateRenderMode;
-        }
-
-        if (NameSelectors is null) return BitNavItemTemplateRenderMode.Normal;
-
-        if (NameSelectors.TemplateRenderMode.Selector is not null)
-        {
-            return NameSelectors.TemplateRenderMode.Selector!(item) ?? BitNavItemTemplateRenderMode.Normal;
-        }
-
-        return item.GetValueFromProperty(NameSelectors.TemplateRenderMode.Name, BitNavItemTemplateRenderMode.Normal);
-    }
-
-    internal string? GetText(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Text;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Text;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Text.Selector is not null)
-        {
-            return NameSelectors.Text.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Text.Name);
-    }
-
-    internal string? GetTitle(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Title;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Title;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Title.Selector is not null)
-        {
-            return NameSelectors.Title.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Title.Name);
-    }
-
-    internal string? GetUrl(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.Url;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.Url;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.Url.Selector is not null)
-        {
-            return NameSelectors.Url.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<string?>(NameSelectors.Url.Name);
-    }
-
-    internal IEnumerable<string>? GetAdditionalUrls(TItem item)
-    {
-        if (item is BitNavItem navItem)
-        {
-            return navItem.AdditionalUrls;
-        }
-
-        if (item is BitNavOption navOption)
-        {
-            return navOption.AdditionalUrls;
-        }
-
-        if (NameSelectors is null) return null;
-
-        if (NameSelectors.AdditionalUrls.Selector is not null)
-        {
-            return NameSelectors.AdditionalUrls.Selector!(item);
-        }
-
-        return item.GetValueFromProperty<IEnumerable<string>?>(NameSelectors.AdditionalUrls.Name);
-    }
-
     internal string GetItemKey(TItem item, string defaultKey)
     {
         return GetKey(item) ?? $"{UniqueId}-{defaultKey}";
@@ -910,16 +240,58 @@ public partial class BitNav<TItem> : BitComponentBase where TItem : class
 
     internal void SetSelectedItemByCurrentUrl()
     {
-        var currentUrl = _navigationManager.Uri.Replace(_navigationManager.BaseUri, "/", StringComparison.Ordinal);
-        var currentItem = Flatten(_items).FirstOrDefault(item => GetUrl(item) == currentUrl
-                                                        || (GetAdditionalUrls(item)?.Contains(currentUrl) ?? false));
+        if (Mode is not BitNavMode.Automatic) return;
+
+        string currentUrl = _navigationManager.Uri.Replace(_navigationManager.BaseUri, "/", StringComparison.Ordinal);
+
+        var currentItem = Flatten(_items).FirstOrDefault(item =>
+        {
+            var match = GetMatch(item) ?? Match ?? BitNavMatch.Exact;
+
+            if (IsMatch(GetUrl(item), match)) return true;
+
+            return GetAdditionalUrls(item)?.Any(u => IsMatch(u, match)) is true;
+        });
 
         _ = SetSelectedItem(currentItem);
+
+        const string DOUBLE_STAR_PLACEHOLDER = "___BIT_NAV_DOUBLESTAR_PLACEHOLDER___";
+        bool IsMatch(string? itemUrl, BitNavMatch? match)
+        {
+            if (itemUrl is null) return false;
+
+            return match switch
+            {
+                BitNavMatch.Exact => itemUrl == currentUrl,
+                BitNavMatch.Prefix => currentUrl.StartsWith(itemUrl, StringComparison.Ordinal),
+                BitNavMatch.Regex => Regex.IsMatch(currentUrl, itemUrl),
+                BitNavMatch.Wildcard => IsWildcardMatch(currentUrl, itemUrl),
+                _ => itemUrl == currentUrl,
+            };
+
+            bool IsWildcardMatch(string input, string pattern)
+            {
+                string regexPattern = $"^{WildcardToRegex(pattern)}$";
+                return Regex.IsMatch(input, regexPattern);
+            }
+
+            string WildcardToRegex(string pattern)
+            {
+                pattern = Regex.Escape(pattern);
+
+                pattern = pattern.Replace(@"\*\*", DOUBLE_STAR_PLACEHOLDER);
+                pattern = pattern.Replace(@"\*", "[^/]*");
+                pattern = pattern.Replace(@"\?", "[^/]");
+                pattern = pattern.Replace(DOUBLE_STAR_PLACEHOLDER, ".*");
+
+                return pattern;
+            }
+        }
     }
 
 
 
-    private List<TItem> Flatten(IList<TItem> e) => e.SelectMany(c => Flatten(GetChildItems(c))).Concat(e).ToList();
+    private List<TItem> Flatten(IList<TItem> e) => [.. e.SelectMany(c => Flatten(GetChildItems(c))), .. e];
 
     private void OnLocationChanged(object? sender, LocationChangedEventArgs args)
     {
